refactor reminder classes
