+++ byte_bot/services/database_service.py
@@ -11,25 +11,6 @@ class UserRecord:
     leetcode_username: str | None = None
 
 
-@dataclass
-class ReminderChannel:
-    id: int
-    name: str
-
-
-@dataclass
-class Reminder:
-    id: int
-    event_id: int
-    event_name: str
-    url: str
-    text: str
-    channel_id: int
-    reminder_minutes: int
-    event_start: int
-    sent_at: int
-
-
 class DatabaseService:
     def __init__(self, database_path: str):
         self.database_path = database_path
@@ -169,231 +150,3 @@ class DatabaseService:
             discord_username=row["discord_username"],
             leetcode_username=row["leetcode_username"],
         )
-
-    def get_reminder_channels(self) -> list[ReminderChannel]:
-        """Return a list of all channels for reminder events."""
-        channels = []
-        with self.get_connection() as connection:
-            rows = connection.execute("SELECT channel_id, name FROM reminder_channels").fetchall()
-
-            if not rows:
-                return channels
-
-            for row in rows:
-                channels.append(ReminderChannel(id=row[0], name=row[1]))
-
-        return channels
-
-    def set_channel_reminder(self, channel_id, name) -> ReminderChannel:
-        """Add or update a channel used for event reminders."""
-        with self.get_connection() as connection:
-            with connection:
-                cursor = connection.execute(
-                    """
-                    INSERT INTO reminder_channels (channel_id, name)
-                    VALUES (?, ?)
-                    ON CONFLICT(channel_id) DO UPDATE SET name = excluded.name
-                    """,
-                    (channel_id, name),
-                )
-                row = cursor.fetchone()
-
-        return ReminderChannel(
-            id=row["channel_id"],
-            name=row["name"],
-        )
-
-    def remove_reminder_channel(self, channel_id) -> bool:
-        """Remove a channel from the reminder channel table given the id."""
-        with self.get_connection() as connection:
-            with connection:
-                cursor = connection.execute(
-                    """
-                        DELETE FROM reminder_channels
-                        WHERE channel_id = ?
-                        """,
-                    (channel_id,),
-                )
-        
-        return cursor.rowcount > 0
-
-    def get_reminders(self) -> list[Reminder]:
-        """Return a list of reminders scheduled to be sent."""
-        reminders = []
-
-        with self.get_connection() as connection:
-            rows = connection.execute(
-                """
-                SELECT id, event_id, event_name, url, text, channel_id, reminder_minutes, event_start
-                FROM reminders
-                WHERE sent_at IS NULL;"
-                """
-            ).fetchall()
-
-            if not rows:
-                return reminders
-
-            for row in rows:
-                reminders.append(
-                    Reminder(
-                        id=row["id"],
-                        event_id=row["event_id"],
-                        event_name=row["event_name"],
-                        url=row["url"],
-                        text=row["text"],
-                        channel_id=row["channel_id"],
-                        reminder_minutes=row["reminder_minutes"],
-                        event_start=row["event_start"],
-                    )
-                )
-
-        return reminders
-
-    def create_reminder(
-        self,
-        event_id: int,
-        event_name: str,
-        url: str,
-        text: str,
-        channel_id: int,
-        reminder_minutes: int,
-        event_start: int,
-    ) -> Reminder | None:
-        """Create and return a new envent reminder and return de updated reminder."""
-
-        with self.get_connection() as connection:
-            with connection:
-                cursor = connection.execute(
-                    """
-                    INSERT INTO reminders (
-                        event_id,
-                        event_name,
-                        url,
-                        text,
-                        channel_id,
-                        reminder_minutes,
-                        event_start
-                    )
-                    VALUES (?, ?, ?, ?, ?, ?, ?)
-                    RETURNING id, event_id, event_name, url, text,
-                            channel_id, reminder_minutes, event_start
-                    """,
-                    (
-                        event_id,
-                        event_name,
-                        url,
-                        text,
-                        channel_id,
-                        reminder_minutes,
-                        event_start,
-                    ),
-                )
-
-                row = cursor.fetchone()
-
-        if row is None:
-            return None
-        
-        return Reminder(
-            id=row["id"],
-            event_id=row["event_id"],
-            event_name=row["event_name"],
-            url=row["url"],
-            text=row["text"],
-            channel_id=row["channel_id"],
-            reminder_minutes=row["reminder_minutes"],
-            event_start=row["event_start"],
-        )
-        
-
-    def mark_reminder_sent(
-        self,
-        reminder_id: int,
-        timestamp: int,
-    ) -> Reminder | None:
-        """Mark a reminder as sent and return the updated reminder."""
-
-        with self.get_connection() as connection:
-            with connection:
-                cursor = connection.execute(
-                    """
-                        UPDATE reminders 
-                        SET sent_at = ?
-                        WHERE id = ?
-                        RETURNING
-                            id,
-                            event_id,
-                            event_name,
-                            url,
-                            text,
-                            channel_id,
-                            reminder_minutes,
-                            event_start,
-                            sent_at
-                        """,
-                    (
-                        timestamp,
-                        reminder_id,
-                    ),
-                )
-                row = cursor.fetchone()
-
-        if row is None:
-            return None
-
-        return Reminder(
-            id=row["id"],
-            event_id=row["event_id"],
-            event_name=row["event_name"],
-            url=row["url"],
-            text=row["text"],
-            channel_id=row["channel_id"],
-            reminder_minutes=row["reminder_minutes"],
-            event_start=row["event_start"],
-            sent_at=row["sent_at"],
-        )
-
-    def update_reminder(self, event_id, timestamp):
-        """Update unsent reminders for an event."""
-        
-        with self.get_connection as connection:
-            with connection:
-                cursor = connection.execute(
-                    """
-                    UPDATE reminders
-                    SET event_start = ?
-                    WHERE event_id = ?
-                        AND sent_at IS NULL
-                    RETURNING
-                        id,
-                        event_id,
-                        event_name,
-                        url,
-                        text,
-                        channel_id,
-                        reminder_minutes,
-                        event_start,
-                        sent_at
-                    """,
-                    (
-                        timestamp,
-                        event_id,
-                    ),
-                )
-
-                rows = cursor.fetchall()
-        
-        return [
-            Reminder(
-                id=row["id"],
-                event_id=row["event_id"],
-                event_name=row["event_name"],
-                url=row["url"],
-                text=row["text"],
-                channel_id=row["channel_id"],
-                reminder_minutes=row["reminder_minutes"],
-                event_start=row["event_start"],
-                sent_at=row["sent_at"],
-            )
-            for row in rows
-        ]
